Estoque: refactor models/service and add Index
Rename EstoqueModel.ProdutosModel -> Produtos and change RegistroProdutosModel totals from double to decimal. Update EstoqueService to use the new property, add stock validation in BaixarEstoque, refactor aggregation/query logic to produce RegistroProdutosModel instances, and adjust persistence flow. Add Index action in EstoqueController to return registros and update site layout navigation to include Produtos and Estoque.

+++ LojaProdutos/Models/RegistroProdutosModel.cs
@@ -4,8 +4,8 @@
     {
         public int ProdutoId { get; set; }
         public string CategoriaNome { get; set; }
-        public double Total { get; set; }
+        public decimal Total { get; set; }
         public DateTime DataCompra { get; set; }
-        public double TotalGeral { get; set; }
+        public decimal TotalGeral { get; set; }
     }
 }
